backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-dotenv
pyyaml
python-multipart
aiofiles
httpx
python-jose[cryptography]
passlib
bcrypt==3.2.2
openai
google-generativeai
anthropic
groq
structlog
watchdog
pytest
pytest-asyncio
psycopg2-binary
openrouter
asyncpg
aiosqlite
email-validator
